Move the camera functions to Graphics
