+++ Sprocket/CMakeLists.txt
@@ -42,17 +42,17 @@ add_library(Sprocket STATIC
             UI/DevUI.cpp
             UI/ImGuiXtra.cpp
 
-            Scene/camera.cpp
             Scene/scene.cpp
             Scene/loader.cpp
             Scene/lua_ecs.cpp
-
+            
             Scene/Systems/basic_systems.cpp
             Scene/Systems/input_system.cpp
-
+            
             Graphics/Animation.cpp
             Graphics/buffer_element_types.cpp
             Graphics/buffer.cpp
+            Graphics/camera.cpp
             Graphics/CubeMap.cpp
             Graphics/frame_buffer.cpp
             Graphics/material.cpp
