+++ Sprocket/Graphics/Rendering/ColliderRenderer.cpp
@@ -1,8 +1,8 @@
 #include "ColliderRenderer.h"
 
-#include <Sprocket/Graphics/render_context.h>
+#include <Sprocket/Graphics/camera.h>
 #include <Sprocket/Graphics/open_gl.h>
-#include <Sprocket/Scene/camera.h>
+#include <Sprocket/Graphics/render_context.h>
 #include <Sprocket/Utility/Maths.h>
 
 #include <glad/glad.h>
