+++ Sprocket/Graphics/Rendering/Scene3DRenderer.cpp
@@ -2,9 +2,9 @@
 
 #include <Sprocket/Graphics/asset_manager.h>
 #include <Sprocket/Graphics/buffer.h>
+#include <Sprocket/Graphics/camera.h>
 #include <Sprocket/Graphics/open_gl.h>
 #include <Sprocket/Graphics/render_context.h>
-#include <Sprocket/Scene/camera.h>
 #include <Sprocket/Utility/Hashing.h>
 #include <Sprocket/Utility/Maths.h>
 #include <Sprocket/Utility/views.h>
