+++ Sprocket/Utility/Maths.h
@@ -24,4 +24,5 @@ glm::vec3 ApplyTransform(const glm::mat4& matrix, const glm::vec3& v);
 glm::vec3 GetMouseRay(const glm::vec2& mousePos, float w, float h, const glm::mat4& view, const glm::mat4& proj);
 
 }
+
 }
